renaming for https://github.com/canjs/can-define-stream/issues/17
