+++ can-stream-kefir.js
