+++ test.html
@@ -1,3 +1,3 @@
-<title>can-stream</title>
-<script src="./node_modules/steal/steal.js" main="can-stream/can-stream_test"></script>
+<title>can-stream-kefir</title>
+<script src="./node_modules/steal/steal.js" main="can-stream-kefir/can-stream-kefir_test"></script>
 <div id="qunit-fixture"></div>
